refactor: copy and content pkg
