+++ oras/provider.py
@@ -9,15 +9,12 @@ import json
 import os
 import shutil
 import sys
-import tempfile
 import urllib
 from contextlib import contextmanager, nullcontext
 from dataclasses import asdict
 from http.cookiejar import DefaultCookiePolicy
 from tempfile import TemporaryDirectory
 from typing import (
-    TYPE_CHECKING,
-    BinaryIO,
     Callable,
     Generator,
     List,
@@ -37,30 +34,14 @@ import oras.main.login as login
 import oras.oci
 import oras.schemas
 import oras.utils
-from oras.layout import Layout
-from oras.copy.descriptor import is_manifest
+from oras.content.layout import LayoutTarget
+from oras.content.registry import RegistryTarget
 from oras.copy import copy as copy_fn
 from oras.copy.options import CopyOptions
 from oras.layout import Layout
-from oras.layout.layout import LayoutTarget
 from oras.logger import logger
 from oras.types import Descriptor, container_type
 from oras.utils.fileio import PathAndOptionalContent
-
-if TYPE_CHECKING:
-    from oras.layout.layout import LayoutTarget
-    from oras.copy.options import CopyOptions
-
-
-# Broad Accept header covering all manifest media types for RegistryTarget.resolve()
-_ACCEPT_ALL_MANIFESTS = ", ".join(
-    [
-        "application/vnd.oci.image.manifest.v1+json",
-        "application/vnd.oci.image.index.v1+json",
-        "application/vnd.docker.distribution.manifest.v2+json",
-        "application/vnd.docker.distribution.manifest.list.v2+json",
-    ]
-)
 
 
 
@@ -1168,184 +1149,3 @@ class Registry:
             response = send()
 
         return response
-
-
-class RegistryTarget:
-    """
-    Adapts a :class:`Registry` + container string into the copy engine's
-    Target, ReferencePusher, and Mounter protocols.
-
-    All operations are scoped to the single repository identified
-    by the container string passed at construction time.
-    """
-
-    def __init__(
-        self,
-        registry: Registry,
-        container: str,
-        opts: "Optional[CopyOptions]" = None,
-    ):
-        self._registry = registry
-        self._container = registry.get_container(container)
-        self._registry.auth.load_configs(self._container)
-        self._opts = opts
-        # Records the most recent manifest PUT response (from push /
-        # push_reference) so callers like Registry.push can return the real
-        # upload response rather than issuing a follow-up GET.
-        self.last_manifest_response: "Optional[requests.Response]" = None
-
-    def _manifest_url(self, ref: str) -> str:
-        """Build full manifest URL for a reference (tag or digest)."""
-        return f"{self._registry.prefix}://{self._container.manifest_url(ref)}"
-
-    def _upload_blob_url(self) -> str:
-        """Build full upload blob URL."""
-        return f"{self._registry.prefix}://{self._container.upload_blob_url()}"
-
-    def _put_manifest(
-        self, reference: str, data: bytes, media_type: str, record: bool
-    ) -> requests.Response:
-        """PUT manifest bytes to a reference (tag or digest).
-
-        Shared by :meth:`push` (manifest branch), :meth:`tag`, and
-        :meth:`push_reference`. When ``record`` is True the response is stored
-        on ``last_manifest_response`` so callers like ``Registry.push`` can
-        return the real upload response instead of issuing a follow-up GET.
-        """
-        url = self._manifest_url(reference)
-        headers = {"Content-Type": media_type}
-        response = self._registry.do_request(
-            url, "PUT", data=data, headers=headers
-        )
-        if record:
-            self.last_manifest_response = response
-        self._registry._check_200_response(response)
-        return response
-
-    def _head_manifest(self, reference: str) -> requests.Response:
-        """Issue a HEAD for a manifest reference with a broad Accept header."""
-        url = self._manifest_url(reference)
-        headers = {"Accept": _ACCEPT_ALL_MANIFESTS}
-        return self._registry.do_request(url, "HEAD", headers=headers)
-
-    def fetch(self, desc: Descriptor) -> BinaryIO:
-        """Fetch content for a descriptor. Routes blob vs manifest."""
-        if is_manifest(desc):
-            url = self._manifest_url(desc["digest"])
-            headers = {"Accept": desc.get("mediaType", "")}
-            response = self._registry.do_request(url, "GET", headers=headers)
-            self._registry._check_200_response(response)
-            return io.BytesIO(response.content)
-        else:
-            response = self._registry.get_blob(self._container, desc["digest"])
-            self._registry._check_200_response(response)
-            return io.BytesIO(response.content)
-
-    def exists(self, desc: Descriptor) -> bool:
-        """Check if content exists. Routes blob vs manifest."""
-        if is_manifest(desc):
-            return self._head_manifest(desc["digest"]).status_code == 200
-        return self._registry.blob_exists(desc, self._container)
-
-    def push(self, desc: Descriptor, content: BinaryIO) -> None:
-        """Push content for a descriptor. Routes blob vs manifest."""
-        if is_manifest(desc):
-            data = content.read()
-            self._put_manifest(
-                desc["digest"], data, desc.get("mediaType", ""), record=True
-            )
-        else:
-            tmp = None
-            try:
-                tmp = tempfile.NamedTemporaryFile(delete=False)
-                # Stream rather than buffer: large layers (and chunked
-                # uploads) must not be fully materialized in memory.
-                shutil.copyfileobj(content, tmp)
-                tmp.close()
-                do_chunked = False
-                chunk_size = oras.defaults.default_chunksize
-                if self._opts is not None:
-                    do_chunked = self._opts.graph.do_chunked
-                    chunk_size = (
-                        self._opts.graph.chunk_size
-                        or oras.defaults.default_chunksize
-                    )
-                self._registry.upload_blob(
-                    tmp.name,
-                    self._container,
-                    desc,
-                    do_chunked=do_chunked,
-                    chunk_size=chunk_size,
-                )
-            finally:
-                if tmp is not None:
-                    try:
-                        os.unlink(tmp.name)
-                    except OSError:
-                        pass
-
-    def tag(self, desc: Descriptor, reference: str) -> None:
-        """Tag a descriptor with a reference by re-uploading the manifest."""
-        data = self.fetch(desc).read()
-        self._put_manifest(
-            reference, data, desc.get("mediaType", ""), record=False
-        )
-
-    def resolve(self, reference: str) -> Descriptor:
-        """Resolve a reference to a descriptor via HEAD request."""
-        response = self._head_manifest(reference)
-        self._registry._check_200_response(response)
-        return {
-            "mediaType": response.headers.get("Content-Type", ""),
-            "digest": response.headers.get("Docker-Content-Digest", ""),
-            "size": int(response.headers.get("Content-Length", 0)),
-        }
-
-    def push_reference(
-        self, desc: Descriptor, content: BinaryIO, reference: str
-    ) -> None:
-        """Atomic push + tag: PUT manifest bytes to the reference URL."""
-        data = content.read()
-        self._put_manifest(
-            reference, data, desc.get("mediaType", ""), record=True
-        )
-
-    def mount(
-        self,
-        desc: Descriptor,
-        from_repo: str,
-        get_content: Callable[[], BinaryIO],
-    ) -> None:
-        """
-        Attempt cross-repo blob mount, falling back to regular upload.
-
-        POSTs to the blob upload endpoint with mount and from params.
-        If the registry returns 201, the mount succeeded. If 202, the
-        mount failed and we complete the upload session with get_content().
-        """
-        url = oras.utils.append_url_params(
-            self._upload_blob_url(),
-            {"mount": desc["digest"], "from": from_repo},
-        )
-        response = self._registry.do_request(url, "POST")
-        if response.status_code == 201:
-            return  # Mount succeeded
-
-        # Mount failed (202) — fall back to regular upload
-        content = get_content()
-        data = content.read()
-        session_url = self._registry._get_location(response, self._container)
-        if not session_url:
-            raise ValueError("Mount fallback: no session URL in response")
-
-        blob_url = oras.utils.append_url_params(
-            session_url, {"digest": desc["digest"]}
-        )
-        headers = {
-            "Content-Length": str(len(data)),
-            "Content-Type": "application/octet-stream",
-        }
-        response = self._registry.do_request(
-            blob_url, "PUT", data=data, headers=headers
-        )
-        self._registry._check_200_response(response)
